+++ CHANGELOG.md
@@ -17,6 +17,7 @@ NetBox now makes use of its own `Tag` model instead of the vanilla model which s
 ## Enhancements
 
 * [#2324](https://github.com/digitalocean/netbox/issues/2324) - Add color option for tags
+* [#2643](https://github.com/digitalocean/netbox/issues/2643) - Add `description` field to console/power components and device bays
 * [#2791](https://github.com/digitalocean/netbox/issues/2791) - Add a comment field for tags
 * [#2926](https://github.com/digitalocean/netbox/issues/2926) - Add changelog to the Tag model